02_Industrial_Defect_Detection/requirements.txt
streamlit
tensorflow
gdown
numpy
pillow
matplotlib